wesave/requirements.txt
alembic==1.14.0
Authlib==1.3.2
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
Flask-Assets==2.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpickle==3.4.2
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
python-digitalocean==1.17.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
stripe==11.2.0
tailwind==3.1.5b0
typing_extensions==4.12.2
urllib3==2.2.3
webassets==2.0
Werkzeug==3.1.3
